ore: remove invalid i32 -> usize CastFrom impl
This simply was not a valid CastFrom implementation. It is not in
general possible to cast an i32 to a usize losslessly, because the i32
might be negative.

This code also fixes up a number of places where we were relying on this
incorrect cast.

+++ src/testdrive/src/action.rs
@@ -95,7 +95,7 @@ pub struct State {
     kafka_admin_opts: rdkafka::admin::AdminOptions,
     kafka_config: ClientConfig,
     kafka_producer: rdkafka::producer::FutureProducer<rdkafka::client::DefaultClientContext>,
-    kafka_topics: HashMap<String, i32>,
+    kafka_topics: HashMap<String, usize>,
     aws_region: rusoto_core::Region,
     aws_account: String,
     aws_credentials: AwsCredentials,
